+++ packages/semantic-release-pnpm/__tests__/intigration/semantic-release-integration.test.ts
@@ -49,7 +49,7 @@ describe("semantic-release-integration", () => {
         await stop();
     });
 
-    it('should skip npm auth verification if "npmPublish" is false', async () => {
+    it("should skip npm auth verification if \"npmPublish\" is false", async () => {
         expect.assertions(1);
 
         await writeJson(join(cwd, "package.json"), {
@@ -74,7 +74,7 @@ describe("semantic-release-integration", () => {
         ).resolves.not.toThrow();
     });
 
-    it('should skip npm auth verification if "package.private" is true', async () => {
+    it("should skip npm auth verification if \"package.private\" is true", async () => {
         expect.assertions(1);
 
         await writeJson(join(cwd, "package.json"), {
@@ -100,7 +100,7 @@ describe("semantic-release-integration", () => {
         ).resolves.not.toThrow();
     });
 
-    it('should skip npm token verification if "package.private" is true', async () => {
+    it("should skip npm token verification if \"package.private\" is true", async () => {
         expect.assertions(1);
 
         await writeJson(join(cwd, "package.json"), {